Better log messages

+++ core/src/main/java/net/lugocorp/kingdom/utils/math/Coords.java
@@ -62,7 +62,7 @@ public class Coords {
          * Returns the euclidean distance between the two grid Points
          */
         public float distance(Point p1, Point p2) {
-            return Math.sqrt(Math.pow(p1.x - p2.x, 2) + Math.pow(p1.y - p2.y, 2));
+            return (float) Math.sqrt(Math.pow(p1.x - p2.x, 2) + Math.pow(p1.y - p2.y, 2));
         }
     }
